move tictactoe/index.ts -> hooks/useTicTacToe.ts
